mercor/requirements.txt
pyairtable==2.0.0
openai==1.3.7
python-dotenv==1.0.0
